Rearrange logic to DRY up code
